+++ CHANGELOG.md
@@ -9,7 +9,8 @@ CHANGELOG
 
 ## Upcoming version ##
 
-- 🌟 Add cronjob to automatically close stations when the prices are no longer updated
+- 🌟 Add price storing for price history ([#17](https://github.com/tankste/backend/issues/17))
+- 🌟 Add cronjob to automatically close stations when the prices are no longer updated ([#15](https://github.com/tankste/backend/issues/15))
 
 ## 1.2.1 (2025-01-03) ##